requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
PyYAML>=6.0
torch>=2.2
tqdm>=4.66
fonttools>=4.50
